A lot of changes, see description for info
Added new drops to the world evil bosses and made it drop both from their bags or on kill.
Added two new debuffs.
Added an unfinished camera player which allows manipulation of the camera.
Added 2 new darts.
Added 3 new materials.
Added an unobtainable ore.
Added 2 new boomerangs.
Added 2 new grenades.
Added new textures for new items.
Added a new critter, small wyvern.
Added a lot of comments for TODO's.
Added a lot of documentation for my methods.
Added ModBoomerang and ModGrenade.

Made every item, debuff and projectile have their names in their file.
Made Sadists Pearl unobtainable.

Removed Eye of Cthulhu rework until I fully make it.

Changed recipes of a few items.
Updated localization.

+++ Core/Net/Handlers/HandlerType.cs
@@ -1,6 +1,6 @@
 ﻿namespace KawaggyMod.Core.Net.Handlers
 {
-    public enum HandlerType : byte
+    public enum HandlerType : uint
     {
         KawaggyPlayerHandler,
         SummonPlayerHandler
